task-24-replace mock data and add leaderboard to router
